chore: migrate to use jest for testing
