+++ package.json
@@ -8,7 +8,7 @@
   ],
   "scripts": {
     "build": "tsc",
-    "test": "ts-node ./tests/test.ts"
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -28,10 +28,12 @@
     "access": "public"
   },
   "devDependencies": {
+    "@types/jest": "^29.4.0",
     "@types/node": "^16",
     "@typescript-eslint/parser": "^5.50.0",
     "@typescript-eslint/utils": "^5.50.0",
-    "ts-node": "^10.9.1",
+    "jest": "^29.4.2",
+    "ts-jest": "^29.0.5",
     "typescript": "^4.9.5"
   },
   "peerDependencies": {
