+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
   "include": ["src"],
+  "exclude": ["**/*.spec.ts"],
   "compilerOptions": {
     /* Visit https://aka.ms/tsconfig to read more about this file */
